Rename github to github_tools so that modules don't conflict with PyGithub
