+++ github_tools/github_helpers_test.py
@@ -9,7 +9,7 @@ SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 import unittest
 from unittest import mock
 
-import github_helpers
+from github_tools import github_helpers
 
 
 _TEST_QUERY = """
