+++ github_tools/update_label_access.py
@@ -14,12 +14,19 @@ SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 """
 
 import argparse
+import os
+import sys
 
 # https://github.com/PyGithub/PyGithub
 # GraphQL is preferred, but falling back to pygithub for unsupported mutations.
 import github
 
-import github_helpers
+# To support direct runs, ensure the pythonpath has the repo root.
+_PYTHONPATH = os.path.realpath(os.path.join(os.path.dirname(__file__), ".."))
+if _PYTHONPATH not in sys.path:
+    sys.path.insert(0, _PYTHONPATH)
+
+from github_tools import github_helpers
 
 # The organization to mirror members from.
 _ORG = "carbon-language"
